refactor: split amm/helpers.go into 8 cohesive files (#76)
Split the 2,266-line helpers.go into files by concern:
- validate.go: input validation (4 functions)
- asset.go: asset matching and comparison (13 functions)
- keylet.go: address/keylet computation (7 functions)
- math.go: arithmetic, conversion, LP token calculation (21 functions)
- formulas.go: trading fees, pool equations, rounding (22 functions)
- serialize.go: AMMData serialization/deserialization (10 functions)
- trustline.go: trust line mutations (4 functions + 1 type)
- pool.go: pool balance queries (5 functions)

Pure file reorganization — no logic changes.
